Add files via upload
Here is a fully updated version of mine. There is a bug apparently in the "if (!isValid(key))" condition and I can't seem to figure it out. It's the one small thing that has been bugging me all day. I think the isValid function should return true, because after all if there is nothing wrong with a valid key then it is true right? Other than that, all good